Move useragent info to ServiceVersionInfo
