+++ codegen/src/main/java/software/amazon/awssdk/codegen/poet/PoetExtension.java
@@ -83,10 +83,6 @@ public class PoetExtension {
         return ClassName.get(model.getMetadata().getFullClientInternalPackageName(), "ServiceVersionInfo");
     }
 
-    public ClassName getServiceVersionUserAgentClass() {
-        return ClassName.get(model.getMetadata().getFullClientInternalPackageName(), "ServiceVersionUserAgent");
-    }
-
     public ClassName getEnvironmentTokenSystemSettingsClass() {
         return ClassName.get(model.getMetadata().getFullClientInternalPackageName(), "EnvironmentTokenSystemSettings");
     }
