+++ codegen/src/main/java/software/amazon/awssdk/codegen/poet/client/AsyncClientClass.java
@@ -234,7 +234,7 @@ public final class AsyncClientClass extends AsyncClientInterface {
                                       SdkClientOption.class,
                                       SdkClientOption.class,
                                       ClassName.get(model.getMetadata().getFullClientInternalPackageName(),
-                                        "ServiceVersionUserAgent"));
+                                        "ServiceVersionInfo"));
 
         FieldSpec protocolFactoryField = protocolSpec.protocolFactory(model);
         if (model.getMetadata().isJsonProtocol()) {
