+++ codegen/src/main/java/software/amazon/awssdk/codegen/poet/client/specs/ServiceVersionInfoSpec.java
@@ -22,6 +22,7 @@ import com.squareup.javapoet.FieldSpec;
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.TypeSpec;
 import javax.lang.model.element.Modifier;
+import software.amazon.awssdk.annotations.SdkInternalApi;
 import software.amazon.awssdk.codegen.model.intermediate.IntermediateModel;
 import software.amazon.awssdk.codegen.poet.ClassSpec;
 import software.amazon.awssdk.codegen.poet.PoetExtension;
@@ -29,9 +30,11 @@ import software.amazon.awssdk.codegen.poet.PoetUtils;
 
 public class ServiceVersionInfoSpec implements ClassSpec {
     private final PoetExtension poetExtension;
+    private final IntermediateModel model;
 
     public ServiceVersionInfoSpec(IntermediateModel model) {
         this.poetExtension = new PoetExtension(model);
+        this.model = model;
     }
 
     @Override
@@ -39,12 +42,14 @@ public class ServiceVersionInfoSpec implements ClassSpec {
         TypeSpec.Builder builder = TypeSpec.classBuilder("ServiceVersionInfo")
                                            .addModifiers(Modifier.PUBLIC, Modifier.FINAL)
                                            .addAnnotation(PoetUtils.generatedAnnotation())
+                                           .addAnnotation(SdkInternalApi.class)
                                            .addField(FieldSpec.builder(
                                                String.class, "VERSION", Modifier.PUBLIC, Modifier.STATIC, Modifier.FINAL)
                                                               .initializer("$S", SDK_VERSION)
                                                               .addJavadoc("Returns the current version for the AWS SDK in which"
                                                                           + " this class is running.")
                                                               .build())
+                                           .addField(userAgentField())
                                            .addMethod(privateConstructor());
 
         return builder.build();
@@ -54,6 +59,20 @@ public class ServiceVersionInfoSpec implements ClassSpec {
         return MethodSpec.constructorBuilder()
                          .addModifiers(Modifier.PRIVATE)
                          .build();
+    }
+
+    private FieldSpec userAgentField() {
+        return FieldSpec.builder(String.class, "USER_AGENT", Modifier.PUBLIC, Modifier.STATIC, Modifier.FINAL)
+                        .initializer("$S", transformServiceId(model.getMetadata().getServiceId()) + "#" + SDK_VERSION)
+                        .addAnnotation(SdkInternalApi.class)
+                        .addJavadoc("Returns a user agent containing the service and "
+                                    + "version info")
+                        .build();
+    }
+
+    private String transformServiceId(String serviceId) {
+        // According to User Agent 2.0 spec, replace spaces with underscores
+        return serviceId.replace(" ", "_");
     }
 
     @Override
