--- codegen/src/test/resources/software/amazon/awssdk/codegen/poet/client/specs/test-service-version-user-agent-class.java
@@ -1,15 +1,0 @@
-package software.amazon.awssdk.services.json.internal;
-
-import java.lang.String;
-import software.amazon.awssdk.annotations.Generated;
-
-@Generated("software.amazon.awssdk:codegen")
-public final class ServiceVersionUserAgent {
-  /**
-   * Returns a user agent containing the service and version info
-   */
-  public static final String USER_AGENT = "Json_Service#{{VERSION}}";
-
-  private ServiceVersionUserAgent() {
-  }
-}
